Refactor days in month value
