+++ corehq/apps/commtrack/tests/test_settings.py
@@ -1,7 +1,8 @@
 from django.test import TestCase
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.commtrack.models import CommtrackConfig, ConsumptionConfig, StockRestoreConfig
-from corehq.apps.commtrack.tests.util import bootstrap_domain, DAYS_IN_MONTH
+from corehq.apps.commtrack.tests.util import bootstrap_domain
+from corehq.apps.commtrack.const import DAYS_IN_MONTH
 from corehq.apps.consumption.shortcuts import set_default_consumption_for_domain
