+++ corehq/apps/commtrack/tests/test_xml.py
@@ -19,7 +19,8 @@ from corehq.apps.commtrack.tests.util import CommTrackTest, get_ota_balance_xml,
 from casexml.apps.case.tests.util import check_xml_line_by_line, check_user_has_case
 from corehq.apps.hqcase.utils import get_cases_in_domain
 from corehq.apps.receiverwrapper import submit_form_locally
-from corehq.apps.commtrack.tests.util import make_loc, make_supply_point, DAYS_IN_MONTH
+from corehq.apps.commtrack.tests.util import make_loc, make_supply_point
+from corehq.apps.commtrack.const import DAYS_IN_MONTH
 from corehq.apps.commtrack.requisitions import get_notification_message
 from corehq.apps.commtrack.tests.data.balances import (
     balance_ota_block,
