+++ corehq/apps/consumption/const.py
@@ -1,0 +1,1 @@
+DAYS_IN_MONTH = 30.0
